Added the option to enable or disable specific languages for each task.
Previously you can set the list of allowed languages only per contest,
with this patch you can customize the list for each task or choose to
use the contest settings.

The motivation around this feature came from a strange behavior (bug?)
when in a task there is a missing grader. It may happen that in a big
contest with hundreds of tasks (this is out situation), some of them
only supports a limited number of languages (because we have implemented
only a subset of the graders). If someone sends a solution in an allowed
language that has no grader some really bad things happen (workers fail
indefinitely).

This is implemented by adding a new column to the tasks table, if the
value is NULL the default list is used, otherwise this list is used.

All the localization strings has been changed, but in some languages the
translaction has not been updated yet. These languages need a revision:
bg, bs, et, ja, ko, lt, lv, nl, ru, vi, zh_CN, zh_TW.