Config files in YAML http://yaml.org/ (JSON is still supported)
YAML (since 1.2) is a superset of JSON that is way more sutable for
configuration files: it supports comments, has a nicer syntax and it's
less verbose than JSON. Since it's a superset of JSON all the old
configuration files should be supported.

This change does not introduce new dependencies, PyYAML is required by
some importers and installed by default following the docs.